refactor: actions/git.ts を責務別モジュールに分割 (#226)
958行の god-file を関数移動＋barrel 再輸出で分割。振る舞い不変・公開 API 不変・
import パス不変の純移動。安全網は既存 git.test.ts(54テスト)を無改変で全緑。

- git-push.ts(407): pushToGitHub/PushToGitHubOptions/PushTimeoutError/世代ガード
- git-pull.ts(416): pullFromGitHub/runPendingRepoSyncIfIdle
- git-orphan.ts(137): observeOrphanPush(#235 遅延結果観測)
- git-connection.ts(31): handleTestConnection
- git.ts(11): 薄い barrel（consumers/テストの ./actions/git 直パス維持）

循環回避: push→pull 一方向 static import・pull→push は appActions 間接維持。
pushProgressGeneration(#238)は git-push 内に閉じ、background phase(#206)の
フラグ順序を温存。architecture.md 追記。

git.ts: 958 -> 11 行

Refs #226
