+++ src/lib/actions/git-pull.ts
@@ -1,0 +1,416 @@
+import { type Leaf, type StaleCheckResult, buildBlobShaCache } from '../types'
+import type { PullOptions } from '../api'
+import { showPullToast } from '../ui'
+import { showConflictDialog } from './conflict-dialog'
+import {
+  settings,
+  notes,
+  leaves,
+  metadata,
+  isDirty,
+  isPulling,
+  isPushing,
+  isPushingBackground,
+  isStale,
+  lastKnownCommitSha,
+  lastPulledPushCount,
+  archiveNotes,
+  archiveLeaves,
+  leftNote,
+  leftLeaf,
+  rightNote,
+  rightLeaf,
+  getPersistedDirtyFlag,
+  executeStaleCheck,
+  setLastPushedSnapshot,
+  addLeafToBaseline,
+  addNotesToBaseline,
+  refreshDirtyState,
+  leafStatsStore,
+  pullProgressStore,
+  rehydrateForRepo,
+  flushAllEditors,
+} from '../stores'
+import { clearAllData, createBackup, restoreFromBackup, saveNotes, saveLeaves } from '../data'
+import { executePull, translateGitHubMessage, canSync } from '../api'
+// #254: 添付フローの挿入フェーズ待ち（詳細は insert-phase.ts のモジュールコメント）
+import { waitForPendingMediaInserts } from '../api/media/insert-phase'
+import { appState, appActions } from '../app-state.svelte'
+import * as nav from '../navigation'
+import { tick } from 'svelte'
+import { get } from 'svelte/store'
+import { _ } from '../i18n'
+import { runPendingRepoSyncIfIdle as runPendingRepoSyncIfIdleShared } from '../sync/repo-sync-queue'
+
+export async function runPendingRepoSyncIfIdle(): Promise<void> {
+  const hasValidConfig = !!(settings.value.token && settings.value.repoName)
+  await runPendingRepoSyncIfIdleShared(
+    {
+      isPulling: isPulling.value,
+      // #206: 背景 Push 中も busy として扱う
+      isPushing: isPushing.value || isPushingBackground.value,
+      isArchiveLoading: appState.isArchiveLoading,
+    },
+    hasValidConfig,
+    appState.pendingRepoSync,
+    () => {
+      appState.pendingRepoSync = false
+    },
+    async () => {
+      // 同期中にリポ切替された場合、予約pull開始前に新リポへ rehydrate する
+      // （旧 DB に新リポの pull 結果が書かれるのを防ぐ）
+      const pendingRehydrate = appState.pendingRehydrateRepo
+      if (pendingRehydrate) {
+        appState.pendingRehydrateRepo = null
+        try {
+          await rehydrateForRepo(pendingRehydrate)
+        } catch (error) {
+          console.error('Failed to rehydrate stores before queued pull:', error)
+        }
+      }
+      await pullFromGitHub(false)
+    }
+  )
+}
+
+/**
+ * Pull処理の統合関数
+ *
+ * @param isInitialStartup - 起動時の初回 Pull か
+ * @param onCancel - Pull キャンセル時のフォールバック処理
+ * @param precomputedStale - 呼び出し元で既に実行した stale check の結果。
+ *   指定した場合は内部で再取得せずそれを使う（#158: app-state の起動時
+ *   スキップ判定で先行実行した結果を流用し、GitHub Refs API の二重呼び出しを避ける）。
+ *   **注意**: 鮮度は検証しない。**この関数呼び出しの直前**に取得した結果だけを渡すこと。
+ *   古い結果を渡すと、その間にリモートが進んでいても up_to_date と判定してしまう。
+ */
+export async function pullFromGitHub(
+  isInitialStartup = false,
+  onCancel?: () => void | Promise<void>,
+  precomputedStale?: StaleCheckResult
+): Promise<void> {
+  const $_ = get(_)
+
+  // 交通整理: Pull/Push中またはアーカイブロード中は不可
+  if (
+    !canSync(isPulling.value, isPushing.value, isPushingBackground.value).canPull ||
+    appState.isArchiveLoading
+  )
+    return
+
+  // 即座にロック取得（この後の非同期処理中にPushが開始されるのを防止）
+  isPulling.value = true
+  // #147 綻び2: この Pull がリポ切替起因かを、直後にフラグが落ちる前に控える。
+  // repoChangePending は handleSettingsChange の repo 切替検知でのみ true になり、
+  // 通常 pull（F5・同期・deep-link 復元・起動時）では false のため、この控えは
+  // 「リポ切替による pull かどうか」を安全に表す（下のスクロール残骸リセット判定に使う）。
+  // queue 経由の切替（同期ビジー中に切替 → 予約 pull）でも印は引き回される:
+  // handleCloseSettings は pendingRepoSync=true のとき repoChangePending を落とさず
+  // 予約 pull の入口（ここ）まで残す。よって repoChangePending 単体で false 化される
+  // ことはなく（direct pull は入口で消費、queue pull は入口まで保持）、両経路で発火する。
+  const isRepoSwitchPull = appState.repoChangePending
+  // pull開始と同時に「pull予約あり」バッジを落とす（進捗%に交代）
+  appState.repoChangePending = false
+  try {
+    // #254: Push 側と同じ理由の待ち。挿入フェーズ着地前に Pull が走ると、
+    // isDirty が挿入前の状態で評価されて dirty 警告を素通りし、Pull の
+    // クリア＆再取得で挿入テキストが消える（メディア参照の喪失＝孤児化）。
+    await waitForPendingMediaInserts()
+    // #254: 挿入が IME composition 中に着地した場合、エディタの updateListener は
+    // onChange を保留する（#186 と同じ機構）ため、store（leaves.value / isDirty）は
+    // 挿入前のまま。Push 側と同様に flush + tick してから下の isDirty 判定に進まないと、
+    // dirty 警告を素通りして Pull の上書きで挿入テキストが消える。
+    flushAllEditors()
+    await tick()
+    // Staleチェックを先に行う。
+    // リモートに変更がない（= 実際のPullが走らない）場合にまで
+    // 「未保存の変更があります。Pullすると上書きされます」系の警告を出すと、
+    // インポート直後（ローカルはリモートと同一commitのまま、配下データだけ増えた
+    // 状態）の設定クローズなど「ローカル先行」状態でも誤った stale 警告
+    // が出てしまうため（#152）。
+    const staleResult =
+      precomputedStale ?? (await executeStaleCheck(settings.value, lastKnownCommitSha.value))
+
+    // リモートに変更なし & Pull完了済み → 実Pullは走らないので dirty 警告も不要
+    if (staleResult.status === 'up_to_date' && appState.isPullCompleted) {
+      showPullToast($_('github.noRemoteChanges'), 'success')
+      return
+    }
+
+    // 未保存の変更がある場合は確認（PWA強制終了後の再起動も考慮）
+    // ここから先は実際にPullが走る可能性があるパスのみ
+    if (isDirty.value || getPersistedDirtyFlag()) {
+      if (isInitialStartup) {
+        // 起動時: push first は選べない（まだPullしていないため disablePush=true）。
+        // 共通ヘルパー経由で他経路と本文・診断情報を統一する（#201）。
+        // disablePush=true により戻り値から 'push' が型レベルで除外される。
+        // staleResult は precomputedStale 経由で「やや古い」値が来る可能性があるが、
+        // 影響は本文 SHA の表示のみ（判定ロジックには使わない）ため受容範囲。
+        const choice = await showConflictDialog({
+          kind: 'startup-dirty',
+          staleResult,
+          localPushCount: metadata.value.pushCount,
+          settings: settings.value,
+          disablePush: true,
+        })
+        if (choice !== 'pull') {
+          await onCancel?.()
+          return
+        }
+      } else {
+        // 通常時: Push first / Pull (overwrite) / Cancel の3択
+        const choice = await showConflictDialog({
+          kind: 'pull-dirty',
+          staleResult,
+          localPushCount: metadata.value.pushCount,
+          settings: settings.value,
+        })
+
+        if (choice === 'push') {
+          // Push first: isPullingロックを解放してPush→Pull
+          isPulling.value = false
+          await appActions.pushToGitHub()
+          // Push後に再度Pull（再帰呼び出し）
+          return pullFromGitHub(false, onCancel)
+        } else if (choice === 'cancel' || choice === null) {
+          await onCancel?.()
+          return
+        }
+        // choice === 'pull' → 続行（上書き）
+      }
+    }
+
+    switch (staleResult.status) {
+      case 'up_to_date':
+        // リモートに変更なし（初回Pull前のみここに到達。実Pull処理を続行して初期化）
+        break
+
+      case 'stale':
+        // リモートに変更あり → Pull実行
+        console.log(
+          `Pull needed: remote(${staleResult.remoteCommitSha}) !== local(${staleResult.localCommitSha})`
+        )
+        break
+
+      case 'check_failed':
+        // チェック失敗 → Pull実行（サーバー側でエラー表示される）
+        console.warn('Stale check failed, proceeding with pull:', staleResult.reason)
+        break
+    }
+
+    // Pull開始準備
+    appState.isLoadingUI = true
+    appState.isFirstPriorityFetched = false
+    appState.isPullCompleted = false
+
+    // Pull開始を通知
+    showPullToast($_('toast.pullStart'))
+
+    // Pull失敗時のデータ保護: 既存データをバックアップ
+    const backup = await createBackup()
+    const hasBackupData = backup.notes.length > 0 || backup.leaves.length > 0
+    if (hasBackupData) {
+      console.log(
+        `Created backup before Pull (${backup.notes.length} notes, ${backup.leaves.length} leaves)`
+      )
+    }
+
+    // blob SHAキャッシュ用: クリア前にバックアップのリーフからSHA→Leafのマップを構築
+    // dirtyな状態ではキャッシュを使わない（編集後contentがリモート由来として扱われるのを防止）
+    const cachedLeafMap = isDirty.value ? new Map<string, Leaf>() : buildBlobShaCache(backup.leaves)
+
+    // 重要: GitHubが唯一の真実の情報源（Single Source of Truth）
+    await clearAllData()
+    notes.value = []
+    leaves.value = []
+    appState.loadingLeafIds = new Set()
+    appActions.resetLeafStats()
+    leftNote.value = null
+    leftLeaf.value = null
+    rightNote.value = null
+    rightLeaf.value = null
+
+    const options: PullOptions = {
+      cachedLeaves: cachedLeafMap.size > 0 ? cachedLeafMap : undefined,
+      // ノート構造確定時
+      onStructure: (notesFromGitHub, metadataFromGitHub, leafSkeletons) => {
+        notes.value = notesFromGitHub
+        metadata.value = metadataFromGitHub
+        addNotesToBaseline(notesFromGitHub)
+
+        appState.leafSkeletonMap = new Map(leafSkeletons.map((s) => [s.id, s]))
+        appState.loadingLeafIds = new Set(leafSkeletons.map((s) => s.id))
+
+        pullProgressStore.start(leafSkeletons.length)
+
+        return nav.getPriorityFromUrl(notesFromGitHub)
+      },
+
+      // 各リーフ取得完了時
+      onLeaf: (leaf) => {
+        leaves.value = [...leaves.value, leaf]
+        leafStatsStore.addLeaf(leaf.id, leaf.content)
+        addLeafToBaseline(leaf)
+        const currentLoadingIds = appState.loadingLeafIds
+        currentLoadingIds.delete(leaf.id)
+        appState.loadingLeafIds = new Set(currentLoadingIds)
+        pullProgressStore.increment()
+      },
+
+      // 第1優先リーフ取得完了時
+      onPriorityComplete: () => {
+        appState.isFirstPriorityFetched = true
+        appState.isLoadingUI = false
+
+        if (isInitialStartup) {
+          appState.isRestoringFromUrl = true
+          appActions.restoreStateFromUrl(true)
+          appState.isRestoringFromUrl = false
+        } else {
+          appActions.restoreStateFromUrl(false)
+        }
+      },
+    }
+
+    const result = await executePull(settings.value, options)
+
+    if (result.success) {
+      appState.isPullCompleted = true
+      appState.selectedIndexLeft = 0
+      appState.selectedIndexRight = 0
+
+      // #147 綻び2: リポ切替起因の Pull のときだけ、旧リポの一覧で残った
+      // スクロール位置をリセットする。home→home では HomeView が remount されず
+      // .main-pane の scrollTop が残るため。selectedIndex=0 は全 pull 成功で走るが、
+      // scroll リセットは切替時に限定する（通常 pull で発火すると deep-link 復元時に
+      // スクロールが飛ぶデグレになる）。.main-pane の参照は app-state.svelte.ts の
+      // PWA 復帰レイアウト修復と同じ .left-column/.right-column 配下セレクタに合わせる。
+      if (isRepoSwitchPull) {
+        await tick()
+        document
+          .querySelectorAll('.left-column .main-pane, .right-column .main-pane')
+          .forEach((el) => {
+            ;(el as HTMLElement).scrollTop = 0
+          })
+      }
+
+      const currentLeaves = leaves.value
+      const currentLeafMap = new Map(currentLeaves.map((l) => [l.id, l]))
+      const sortedLeaves = result.leaves
+        .sort((a, b) => a.order - b.order)
+        .map((leaf) => {
+          const currentLeaf = currentLeafMap.get(leaf.id)
+          if (currentLeaf && currentLeaf.content !== leaf.content) {
+            // Pull中に編集されたリーフ: contentは編集後を採用し、blobShaはクリア
+            // blobShaを残すと次回Pullでキャッシュヒットし、リモートの内容が取得されない
+            return { ...leaf, content: currentLeaf.content, blobSha: undefined }
+          }
+          return leaf
+        })
+      leaves.value = sortedLeaves
+      appActions.rebuildLeafStats(sortedLeaves, result.notes)
+
+      if (result.commitSha) {
+        lastKnownCommitSha.value = result.commitSha
+      }
+      lastPulledPushCount.value = result.metadata.pushCount
+
+      setLastPushedSnapshot(result.notes, result.leaves, archiveNotes.value, archiveLeaves.value)
+
+      // #207: Pull 成功パスでも save を await する。await しないと、Pull 直後にタブが
+      // 閉じる / リロードされる経路で次回 createBackup() が空 IndexedDB を読み、
+      // blobSha キャッシュが効かなくなる（pullIncomplete 経路と同じクラスの事故）。
+      // 一方が失敗しても他方は続行できるように Promise.allSettled で握る。
+      const [notesResult, leavesResult] = await Promise.allSettled([
+        saveNotes(result.notes),
+        saveLeaves(sortedLeaves),
+      ])
+      if (notesResult.status === 'rejected') {
+        console.error('Failed to persist notes:', notesResult.reason)
+      }
+      if (leavesResult.status === 'rejected') {
+        console.error('Failed to persist leaves:', leavesResult.reason)
+      }
+
+      await tick()
+      refreshDirtyState()
+      isStale.value = false
+    } else {
+      appState.isPullCompleted = false
+      if (result.message === 'github.pullIncomplete') {
+        console.error('Pull incomplete: some leaves failed to fetch. UI remains locked.')
+        appState.isFirstPriorityFetched = false
+
+        // 途中まで取得したリーフをIndexedDBに保存（次回Pullのblobキャッシュ用）
+        // これにより数回のPullで全リーフが揃い、最終的にPullが成功する
+        // #207: await して完了を待ってからこのPullサイクルを閉じる。
+        //   await しないと、次回 Pull の createBackup() が空のIndexedDBを読んでしまい、
+        //   blobSha キャッシュが効かず毎回ゼロから再取得する事故が起きる。
+        const partialLeaves = leaves.value
+        const partialNotes = notes.value
+        if (partialLeaves.length > 0 || partialNotes.length > 0) {
+          if (partialLeaves.length > 0) {
+            console.log(
+              `Saving ${partialLeaves.length} partial leaves to IndexedDB for next pull cache`
+            )
+          }
+          try {
+            await Promise.all([
+              partialLeaves.length > 0 ? saveLeaves(partialLeaves) : Promise.resolve(),
+              partialNotes.length > 0 ? saveNotes(partialNotes) : Promise.resolve(),
+            ])
+          } catch (err) {
+            console.error('Failed to persist partial pull cache:', err)
+            // 救済できなくても続行。次回 Pull で再試行される。
+          }
+        }
+
+        // メモリ上はクリア（UIはガラス状を維持）
+        notes.value = []
+        leaves.value = []
+        // ベースラインもリセット（onLeafで追加された部分リーフが残らないように）
+        setLastPushedSnapshot([], [], archiveNotes.value, archiveLeaves.value)
+      } else if (hasBackupData && !isInitialStartup) {
+        // 非初回Pull失敗: 直前の同期済みデータにリストアする（作業中の状態を保護）
+        console.log('Pull failed, restoring from backup...')
+        try {
+          await restoreFromBackup(backup)
+          notes.value = backup.notes
+          leaves.value = backup.leaves
+          appActions.rebuildLeafStats(backup.leaves, backup.notes)
+          appActions.restoreStateFromUrl(false)
+          appState.isFirstPriorityFetched = true
+        } catch (restoreError) {
+          console.error('Failed to restore from backup:', restoreError)
+        }
+      } else if (hasBackupData && isInitialStartup) {
+        // 初回Pull失敗: メモリにはリストアしない（UIはガラス状を維持）が、
+        // IndexedDBにはバックアップを書き戻す（次回PullのblobSHAキャッシュを保全）
+        console.log(
+          `Initial pull failed, restoring IndexedDB for cache (${backup.notes.length} notes, ${backup.leaves.length} leaves, UI remains locked)`
+        )
+        try {
+          await restoreFromBackup(backup)
+        } catch (err) {
+          console.error('Failed to restore IndexedDB for cache:', err)
+        }
+      }
+      // Offlineリーフのみ操作可能。オンライン復帰で自動リトライする。
+    }
+
+    const translatedMessage = translateGitHubMessage(
+      result.message,
+      $_,
+      result.rateLimitInfo,
+      undefined,
+      result.errorCode,
+      result.httpStatus
+    )
+    showPullToast(translatedMessage, result.variant)
+    appState.isLoadingUI = false
+    pullProgressStore.reset()
+  } finally {
+    isPulling.value = false
+    await runPendingRepoSyncIfIdle()
+  }
+}
